--- a/data/Dockerfiles/archiver/Dockerfile
+++ b/data/Dockerfiles/archiver/Dockerfile
@@ -1,0 +1,3 @@
+FROM alpine:latest
+RUN apk add --no-cache -X http://dl-cdn.alpinelinux.org/alpine/edge/testing \
+	    tar mariadb-client zstd