@hachre 2019-05-11
Do you have any before and after benchmark results?