@wucke13 2019-05-12
@hachre That is hard, as running it the second time may profit from caching. I will try to collect some data using a local test instance, and report back though.