@wucke13 2019-05-14
@hachre I added a few results in the original PR post, I will add more later.

Just thinking: 
Maybe LZ4 would be nice for compression too. It is fairly wide spread, compression is really fast (except when using **H**igh **C**ompression, which will take much longer but get better compression ratios) and decompression is always really fast no matter if HC was used.
If I do backups, I'm ok if they take some time, but if I need them (especially for something like a webserver), I'd want them to extract as fast as possible. Probably no other compression algo is going to beat LZ4 decompression wise.