@wucke13 2019-05-15
After a fresh reboot, with `zstd -T0`:
+ 4:39 (min:sec)
+ 9.5G in total
+ No compression for the `sql` dump
+ CPU usage around 50% on all cores, `zstd` distributed across all cores from

no further rebooting (so caching may have help), with `gzip`, the current way:
+ 21:20 (min:sec)
+ 9.6G in total
+ CPU usage around 35% on all cores, one core >90% with `gzip`

So to me the choice is clear. Let's use `zstd`. On my fairly old machine with low IO the backup is more than four times faster, and the resulting backup is a tad smaller, with neglectable cpu hit.

I did not use `zstd` to compress the `sql` file, because `zstd` is not included in the default image we use. Please review