@wucke13 2019-05-18
@NicoDreamzZ Yes! I haven't thought about that, but yes, it will break old backups. Though converting from old to new is trivial, something like this could already work (run when `$PWD` is like `$BACKUP_LOCATION/mailcow-2019-05-18-13-13-00` aka when you're in the folder of one backup):
```
for file in *.tar.gz
do
  pigz --decompress < ${file} | zstd -T0 > ${file%%.gz}.zst
done
pigz --decompress < backup_mysql.gz > backup_mysql.sql
```