@NicoDreamzZ 2019-05-18
@wucke13 I think it would be nice when the script detects which format the backup is and then restores the correct way. My backups are all saved in a big borg repository and converting all is not possible, and later i maybe forget the snippet you posted :D