@wucke13 2019-05-19
@NicoDreamzZ Pardon, but I oppose. Having some sort of decision logic inside of the backup script will make it even more complicated, and more difficult to verify. 

Why would you want to convert *all* your existing backups? As long as it is trivial to convert a backup to the new format, there is no need to convert *anything* until you eventually happen to need a backup. But of course, this conversion snippet should find it's way into the documentation, if this PR gets merged.

Also: nice to hear that you do use borg :+1: