@ocha 2019-06-09
will the backups still be `--rsyncable`?