@wucke13 2019-06-09
@ocha What do you mean by that? The backups will reside in TAR archives, just as before, the only difference is the compression algorithm used. So if you talk about `rsync`ing these files to some other host, of course that'll work just as before.