@ocha 2019-06-17
@wucke13 current solution is using `--rsyncable` flag on gzip which speeds up rsync and saves on bandwidth when changes are small/additive. How important is that feature for overall backup speed assuming one is syncing backup files to a remote server? Below is the explanation of the flag. 

              While  compressing, synchronize the output occasionally based on
              the input.  This increases size by  less  than  1  percent  most
              cases,  but  means that the rsync(1) program can much more effi‐
              ciently synchronize files compressed  with  this  flag.   gunzip
              cannot  tell  the  difference  between a compressed file created
              with this option, and one created without it.