@wucke13 2019-06-17
@ocha Ah I see. Thanks for the explanation. Reading the man page of `zstd`, I found this:
```
 --rsyncable
              zstd will periodically synchronize the compression state to make
              the compressed file more rsync-friendly. There is  a  negligible
              impact  to  compression ratio, and the faster compression levels
              will see a small compression speed hit. This  feature  does  not
              work  with  --single-thread.  You  probably don't want to use it
              with long range mode, since it will decrease  the  effectiveness
              of the synchronization points, but your milage may vary.
```
So, `zstd` offers it too. I will adjust the PR accordingly.