@andryyy 2019-09-26
I don't know. A container for this is too much.

I plan to create a script to create full clones, that can be updated with a command.