@wucke13 2019-09-26
Ok. Will you incorporate `zstd` for compression or ditch compression at all?