@wucke13 2019-10-12
@andryyy Other than that, would it be feasible to you to have an option in the script to have the container pipe out `tar` archives and send them through any command on the host system? That way one doesn't have to alter the container (as no special tools are needed inside of them) and one can easily change the tool used for compression?