@dragoangel 2019-10-17
I♥️Snapshots.
On CentOS7 I using btrfs with btrbk on systems where data can be added and removed before backup actually was done. Unfortunately RHEL discontinued BTRFS support in 8 release. So I went to LVM and Veeam for Linux Free Agent with LVM snapshots. It has wery good speed & compression level vs well cpu usage. In my case bottleneck is 100mbs network between data center and my backup NAS. And good in this all part with backups monitoring. I have nice mail each day with success or failed status, and all details.