@wucke13 2020-03-14
Closing due to lack of interest from maintainers and totally outdated PR. The issue is still relevant IMO. Hopefully someone else will succeed with this!